backend/requirements.txt
fastapi==0.91.0
uvicorn==0.20.0
python-dotenv==0.21.1
# db
SQLAlchemy==2.0.1
psycopg2-binary==2.9.5
#mongodb
pymongo==4.3.3
## The following requirements were added by pip freeze:
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
Automat==22.10.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
dnspython==2.4.2
exceptiongroup==1.1.3
greenlet==3.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==22.10.0
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==1.10.13
pydantic-core==2.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.24.0
tabulate==0.9.0
twisted==23.8.0
twisted-iocpsupport==1.0.4
typing-extensions==4.8.0
yoyo-migrations==8.2.0
zipp==3.17.0
zope.interface==6.1